+++ src/main.rs
@@ -6,11 +6,9 @@ use server::serve;
 mod address_status;
 mod ctx;
 mod db;
-mod logger;
-#[path = "net/http.rs"]
-mod http;
 #[path = "security/jwt.rs"]
 mod jwt;
+mod logger;
 mod proxy;
 #[path = "security/rate_limiter.rs"]
 mod rate_limiter;
@@ -26,10 +24,10 @@ static GLOBAL: jemallocator::Jemalloc = jemallocator::Jemalloc;
 /// A type alias for a generic error, encompassing any error that implements the `Error` trait,
 /// along with traits for thread-safe error handling (`Send` and `Sync`).
 /// This type is typically used across the application to handle errors uniformly.
-type GenericError = Box<dyn std::error::Error + Send + Sync>;
+pub(crate) type GenericError = Box<dyn std::error::Error + Send + Sync>;
 /// A type alias for a generic result, used throughout the application to encapsulate the
 /// outcome of operations that might fail with a `GenericError`.
-type GenericResult<T> = std::result::Result<T, GenericError>;
+pub(crate) type GenericResult<T> = std::result::Result<T, GenericError>;
 
 #[tokio::main]
 async fn main() -> GenericResult<()> {
