+++ src/net/server.rs
@@ -7,9 +7,9 @@ use hyper::{Body, Request, Response, Server, StatusCode};
 
 use super::{GenericError, GenericResult};
 use crate::ctx::{AppConfig, DEFAULT_PORT};
-use crate::http::{http_handler, response_by_status, X_FORWARDED_FOR};
 use crate::logger::tracked_log;
 use crate::proxy::websocket::{should_upgrade_to_socket_conn, socket_handler};
+use crate::proxy::{http_handler, response_by_status, X_FORWARDED_FOR};
 
 pub(crate) fn is_private_ip(ip: &IpAddr) -> bool {
     match ip {
