+++ src/proxy/mod.rs
@@ -1,15 +1,23 @@
+use crate::address_status::{get_address_status_list, post_address_status};
 use crate::ctx::{AppConfig, GenericResult, ProxyRoute};
+use crate::jwt::{get_cached_token_or_generate_one, JwtClaims};
+use crate::logger::tracked_log;
 use crate::rpc::RpcPayload;
-use hyper::header;
+use crate::server::is_private_ip;
 use hyper::header::{HeaderName, HeaderValue};
+use hyper::{header, HeaderMap, Method};
 use hyper::{Body, Request, Response, StatusCode, Uri};
 use proxy_signature::ProxySign;
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
+use serde_json::json;
 use std::net::SocketAddr;
 
 pub(crate) mod http;
 pub(crate) mod websocket;
+
+pub(crate) const APPLICATION_JSON: &str = "application/json";
+pub(crate) const X_FORWARDED_FOR: &str = "x-forwarded-for";
 
 const X_AUTH_PAYLOAD: &str = "X-Auth-Payload";
 const KEEP_ALIVE: &str = "keep-alive";
@@ -223,4 +231,175 @@ fn remove_hop_by_hop_headers(
     }
 
     Ok(())
+}
+
+async fn get_healthcheck() -> GenericResult<Response<Body>> {
+    let json = json!({
+        "health": "ok",
+    });
+
+    Ok(Response::builder()
+        .status(StatusCode::OK)
+        .header("Access-Control-Allow-Origin", "*")
+        .header("Access-Control-Allow-Headers", "*")
+        .header("Access-Control-Allow-Methods", "POST, GET, OPTIONS")
+        .header(header::CONTENT_TYPE, APPLICATION_JSON)
+        .body(Body::from(json.to_string()))?)
+}
+
+fn handle_preflight() -> GenericResult<Response<Body>> {
+    Ok(Response::builder()
+        .status(StatusCode::NO_CONTENT)
+        .header("Access-Control-Allow-Origin", "*")
+        .header("Access-Control-Allow-Headers", "*")
+        .header("Access-Control-Allow-Methods", "POST, GET, OPTIONS")
+        .body(Body::from(Vec::new()))?)
+}
+
+pub(crate) fn response_by_status(status: StatusCode) -> GenericResult<Response<Body>> {
+    Ok(Response::builder()
+        .status(status)
+        .header("Access-Control-Allow-Origin", "*")
+        .header("Access-Control-Allow-Headers", "*")
+        .header("Access-Control-Allow-Methods", "POST, GET, OPTIONS")
+        .body(Body::from(Vec::new()))?)
+}
+
+pub(crate) async fn insert_jwt_to_http_header(
+    cfg: &AppConfig,
+    headers: &mut HeaderMap<HeaderValue>,
+) -> GenericResult<()> {
+    let claims = &JwtClaims::new(cfg.token_expiration_time());
+    let auth_token = get_cached_token_or_generate_one(cfg, claims).await?;
+    headers.insert(
+        header::AUTHORIZATION,
+        format!("Bearer {}", auth_token).parse()?,
+    );
+
+    Ok(())
+}
+
+pub(crate) async fn http_handler(
+    cfg: &AppConfig,
+    mut req: Request<Body>,
+    remote_addr: SocketAddr,
+) -> GenericResult<Response<Body>> {
+    let req_uri = req.uri().clone();
+
+    let is_private_ip = is_private_ip(&remote_addr.ip());
+
+    if is_private_ip {
+        tracked_log(
+            log::Level::Info,
+            remote_addr.ip(),
+            "**not-available**",
+            req.uri(),
+            "Request received from the same network. Security middlewares will be by-passed.",
+        );
+
+        match (req.method(), req_uri.path()) {
+            (&Method::GET, "/") => return get_healthcheck().await,
+            (&Method::GET, "/address-status") => return get_address_status_list(cfg).await,
+            (&Method::POST, "/address-status") => return post_address_status(cfg, req).await,
+            _ => {}
+        };
+    };
+
+    if req.method() == Method::OPTIONS {
+        return handle_preflight();
+    }
+
+    let proxy_route = match req.method() {
+        &Method::GET => match cfg.get_proxy_route_by_uri(req.uri_mut()) {
+            Some(proxy_route) => proxy_route,
+            None => {
+                tracked_log(
+                    log::Level::Warn,
+                    remote_addr.ip(),
+                    "**not-available**",
+                    req_uri,
+                    "Proxy route not found for GET request, returning 404.",
+                );
+
+                return response_by_status(StatusCode::NOT_FOUND);
+            }
+        },
+        _ => match cfg.get_proxy_route_by_inbound(req.uri().path()) {
+            Some(proxy_route) => proxy_route,
+            None => {
+                tracked_log(
+                    log::Level::Warn,
+                    remote_addr.ip(),
+                    "**not-available**",
+                    req_uri,
+                    "Proxy route not found for non-GET request, returning 404.",
+                );
+                return response_by_status(StatusCode::NOT_FOUND);
+            }
+        },
+    };
+
+    let (req, payload) = match generate_payload_from_req(req, &proxy_route.proxy_type).await {
+        Ok(t) => t,
+        Err(e) => {
+            tracked_log(
+                log::Level::Warn,
+                remote_addr.ip(),
+                "**not-available**",
+                req_uri,
+                format!("Received invalid http payload: {e}, returning 401."),
+            );
+            return response_by_status(StatusCode::UNAUTHORIZED);
+        }
+    };
+
+    tracked_log(
+        log::Level::Info,
+        remote_addr.ip(),
+        &payload.proxy_sign().address,
+        &req_uri,
+        "Request and payload data received.",
+    );
+
+    let x_forwarded_for: HeaderValue = match remote_addr.ip().to_string().parse() {
+        Ok(t) => t,
+        Err(_) => {
+            tracked_log(
+                log::Level::Error,
+                remote_addr.ip(),
+                &payload.proxy_sign().address,
+                &req_uri,
+                "Error type casting of IpAddr into HeaderValue, returning 500.",
+            );
+            return response_by_status(StatusCode::INTERNAL_SERVER_ERROR);
+        }
+    };
+
+    if is_private_ip {
+        return proxy(
+            cfg,
+            req,
+            &remote_addr,
+            payload,
+            x_forwarded_for,
+            proxy_route,
+        )
+        .await;
+    }
+
+    if let Err(status_code) =
+        validation_middleware(cfg, &payload, proxy_route, req.uri(), &remote_addr).await
+    {
+        return response_by_status(status_code);
+    }
+
+    proxy(
+        cfg,
+        req,
+        &remote_addr,
+        payload,
+        x_forwarded_for,
+        proxy_route,
+    )
+    .await
 }
